Move the etl jobs into their own subdirectory
The base directory was a mixture of etl jobs and library functionality;
moving the actual etl jobs into a subdirectory is a bit clearer.
