+++ jobs/webcompat-kb/webcompat_kb/etl/standards_positions.py
@@ -6,10 +6,10 @@ from urllib.parse import urlparse
 
 import pydantic
 
-from .base import Context, EtlJob, dataset_arg
-from .bqhelpers import BigQuery, TableSchema
-from .httphelpers import get_json
-from .projectdata import Project
+from ..base import Context, EtlJob, dataset_arg
+from ..bqhelpers import BigQuery, TableSchema
+from ..httphelpers import get_json
+from ..projectdata import Project
 
 
 class StandardsPosition(pydantic.BaseModel):
